pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "medical-vqa-vlm"
version = "0.1.0"
description = "Domain Adaptation of Lightweight VLMs for Medical VQA"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Taeuk Hwang", email = "taeuk@example.com"},
]

dependencies = [
    # Core ML
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.45.0",
    "accelerate>=0.34.0",
    # QLoRA / PEFT
    "peft>=0.13.0",
    "bitsandbytes>=0.44.0",
    "trl>=0.12.0",
    # Datasets
    "datasets>=3.0.0",
    "Pillow>=10.0.0",
    # Qwen VL utils
    "qwen-vl-utils>=0.0.8",
    # Florence-2 vision tower
    "timm>=1.0.0",
    # Experiment tracking
    "wandb>=0.18.0",
    # HPO
    "optuna>=4.0.0",
    # Evaluation
    "scikit-learn>=1.5.0",
    "scipy>=1.14.0",
    "nltk>=3.9.0",
    "num2words>=0.5.13",
    # Utilities
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "tqdm>=4.66.0",
    "pandas>=2.2.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "einops>=0.8.2",
]

[project.optional-dependencies]
unsloth = [
    "unsloth[cu128-torch2.6]>=2025.3.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu128" }]
torchvision = [{ index = "pytorch-cu128" }]
